requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.8.0rc2
distlib==0.3.8
Django==3.2
django-filter==23.5
django-graphql-jwt==0.4.0
django-graphql-social-auth==0.1.4
django-js-asset==2.1.0
filelock==3.13.1
graphene==3.3
graphene-django==3.1.5
graphene-file-upload==1.3.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.6
oauthlib==3.2.2
Pillow==10.1.0
platformdirs==4.2.0
promise==2.3
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
urllib3==2.2.0
virtualenv==20.25.0
